refactor: rename `transcript_protocol` module to `merlin_transcript_core`
